Separating dev webpack from dev server to use it in local-start.sh
