+++ packages/dashboard-backend/src/index.ts
@@ -76,7 +76,13 @@ server.listen(8080, '0.0.0.0', (err: Error, address: string) => {
     console.error(err);
     process.exit(1);
   }
-  console.log(`Server listening at ${address}`);
+  if (isLocalRun) {
+    // when we're running against keycloak, 0.0.0.0 is not allowed
+    // so suggesting to use whitelisted localhost instead
+    console.log(`Server listening at http://localhost:8080/`);
+  } else {
+    console.log(`Server listening at ${address}`);
+  }
 });
 
 server.ready(() => {
