+++ packages/dashboard-frontend/package.json
@@ -20,8 +20,9 @@
   },
   "scripts": {
     "build": "webpack --config webpack.config.prod.js",
-    "build:source-map": "yarn run compile --devtool source-map",
-    "start": "webpack-dev-server --progress --color --config webpack.config.dev.js",
+    "build:dev": "webpack --progress --color --config webpack.config.dev.js",
+    "build:watch": "yarn build:dev --watch",
+    "start": "webpack-dev-server --progress --color --config webpack.config.dev-server.js",
     "format": "tsfmt -r --useTsfmt tsfmt.json",
     "eslint": "tsc --noEmit && eslint '*/**/*.{js,ts,tsx}'",
     "eslint:fix": "tsc --noEmit && eslint '*/**/*.{js,ts,tsx}' --fix",
